Updated to use new notes create method
git-tfs-id: [http://tfs2010:8080/tfs/gd]$/Delta/Services/AtlantisFramework/4.0;C388271

+++ Providers/Iris/Atlantis.Framework.Providers.Iris/Atlantis.Framework.Providers.Iris.Interface/IIrisProvider.cs
@@ -10,6 +10,6 @@ namespace Atlantis.Framework.Providers.Iris.Interface
       Incident CreateIrisIncident(int subscriberId, string subject, string note, string customerEmailAddress, string ipAddress, string createdBy, int privateLableId, int groupId, int serviceId, string shopperId);
       Note AddIncidentNote(long incidentId, string note, string loginId);
       List<Incident> GetIncidents(string shopperId, DateTime startDate, DateTime endDate, bool deepLoad);
-      List<Note> GetIncidentNotes(long incidentId, long noteId);
+      List<Note> GetIncidentNotes(long incidentId, int noteId);
     }
 }
